Modified variables types (long, int, long long, etc.) to fixed bit type from stdint.h (int32_t, uint32_t, uint64_t, int64_t, etc.)

+++ include/AddonFormat.h
@@ -1,15 +1,15 @@
 
 #ifndef ADDON_FORMAT_H
 #define ADDON_FORMAT_H
+#include <stdint.h>
 #include "Bootil/Bootil.h"
-
 
 namespace Addon
 {
 	static const char*	Ident							= "GMAD";
 	static const char	Version							= 3;
-	static const unsigned int AppID						= 4000;
-	static const unsigned int CompressionSignature		= 0xBEEFCACE;
+	static const uint32_t AppID						= 4000;
+	static const uint32_t CompressionSignature		= 0xBEEFCACE;
 
 
 	struct Header
@@ -21,10 +21,10 @@ namespace Addon
 	struct FileEntry
 	{
 		Bootil::BString	strName;
-		long long		iSize;
-		unsigned long	iCRC;
-		unsigned int	iFileNumber;
-		long long		iOffset;
+		int64_t		iSize;
+		uint32_t	iCRC;
+		uint32_t	iFileNumber;
+		int64_t		iOffset;
 
 		typedef std::list< FileEntry > List;
 	};
@@ -95,9 +95,9 @@ namespace Addon
 	};
 
 	//
-	// This is the position in the file containing a 64 bit unsigned int that represents the file's age
+	// This is the position in the file containing a 64 bit uint32_t that represents the file's age
 	// It's basically the time it was uploaded to Steam - and is set on download/extraction from steam.
 	//
-	static unsigned int			TimestampOffset			= sizeof( Addon::Header ) + sizeof( unsigned long long );
+	static uint32_t			TimestampOffset			= sizeof( Addon::Header ) + sizeof( uint64_t );
 }
 #endif
